creates AddToBucketListButton component with functionality for addToBucketList button from AttractionCard, then renders this component in AttractionCard and AttractionPage
